Split up test helpers header.
